+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Renamed the package `starters` because a package called `projects` was released on CRAN.
 
+* Added name validation (@archetana, #38)
+
 # pRojects 0.0.1.9003
 
 * Added a `NEWS.md` file to track changes to the package.